+++ packages/autopublish/package.js
@@ -3,5 +3,6 @@ Package.describe({
   version: '1.0.3'
 });
 
-// This package is empty; its presence is detected by livedata and
-// accounts-base.
+// This package is empty; its presence is detected by several other packages
+// (such as ddp-server and mongo) which check for the presence of
+// Package.autopublish.