+++ packages/diff-sequence/package.js
@@ -1,0 +1,22 @@
+Package.describe({
+  summary: "An implementation of a diff algorithm on arrays and objects.",
+  version: '1.0.0'
+});
+
+Package.onUse(function (api) {
+  api.export('DiffSequence');
+  api.use(['underscore', 'ejson']);
+  api.addFiles([
+    'diff.js'
+  ]);
+});
+
+Package.onTest(function (api) {
+  api.use('tinytest');
+  api.use('diff-sequence');
+  api.addFiles([
+    'tests.js'
+  ]);
+});
+
+