+++ packages/meteor/package.js
@@ -11,7 +11,7 @@ Package.registerBuildPlugin({
 });
 
 Npm.depends({
-  "double-ended-queue": "2.1.0-0"
+  "meteor-deque": "2.1.0"
 });
 
 Package.onUse(function (api) {