+++ packages/minimongo/id_map.js
@@ -1,6 +1,6 @@
 LocalCollection._IdMap = function () {
   var self = this;
-  IdMap.call(self, LocalCollection._idStringify, LocalCollection._idParse);
+  IdMap.call(self, MongoID.idStringify, MongoID.idParse);
 };
 
 Meteor._inherits(LocalCollection._IdMap, IdMap);