+++ tools/files.js
@@ -15,7 +15,7 @@ var crypto = require('crypto');
 var rimraf = require('rimraf');
 var Future = require('fibers/future');
 var sourcemap = require('source-map');
-var sourcemap_support = require('source-map-support');
+var sourceMapRetrieverStack = require('./source-map-retriever-stack.js');
 
 var utils = require('./utils.js');
 var cleanup = require('./cleanup.js');
@@ -34,21 +34,19 @@ _.extend(files, miniFiles);
 
 var parsedSourceMaps = {};
 var nextStackFilenameCounter = 1;
-var retrieveSourceMap = function (pathForSourceMap) {
-  if (_.has(parsedSourceMaps, pathForSourceMap))
+
+// Use the source maps specified to runJavaScript
+var useParsedSourceMap = function (pathForSourceMap) {
+  // Check our fancy source map data structure, used for isopacks
+  if (_.has(parsedSourceMaps, pathForSourceMap)) {
     return {map: parsedSourceMaps[pathForSourceMap]};
+  }
+
   return null;
 };
 
-sourcemap_support.install({
-  // Use the source maps specified to runJavaScript instead of parsing source
-  // code for them.
-  retrieveSourceMap: retrieveSourceMap,
-  // For now, don't fix the source line in uncaught exceptions, because we
-  // haven't fixed handleUncaughtExceptions in source-map-support to properly
-  // locate the source files.
-  handleUncaughtExceptions: false
-});
+// Try this source map first
+sourceMapRetrieverStack.push(useParsedSourceMap);
 
 // given a predicate function and a starting path, traverse upwards
 // from the path until we find a path that satisfies the predicate.
@@ -288,8 +286,8 @@ var makeTreeReadOnly = function (p) {
     });
   }
   if (stat.isFile()) {
-    var permissions = stat.mode & 0777;
-    var readOnlyPermissions = permissions & 0555;
+    var permissions = stat.mode & 0o777;
+    var readOnlyPermissions = permissions & 0o555;
     if (permissions !== readOnlyPermissions)
       files.chmod(p, readOnlyPermissions);
   }
@@ -356,7 +354,7 @@ files.treeHash = function (root, options) {
       }
       updateHash('file ' + JSON.stringify(relativePath) + ' ' +
                   stat.size + ' ' + files.fileHash(absPath) + '\n');
-      if (stat.mode & 0100) {
+      if (stat.mode & 0o100) {
         updateHash('exec\n');
       }
     } else if (stat.isSymbolicLink()) {
@@ -439,7 +437,7 @@ files.cp_r = function (from, to, options) {
   options = options || {};
 
   var absFrom = files.pathResolve(from);
-  files.mkdir_p(to, 0755);
+  files.mkdir_p(to, 0o755);
 
   _.each(files.readdir(from), function (f) {
     if (_.any(options.ignore || [], function (pattern) {
@@ -465,7 +463,7 @@ files.cp_r = function (from, to, options) {
       // be modified by umask.) We don't copy the mode *directly* because this
       // function is used by 'meteor create' which is copying from the read-only
       // tools tree into a writable app.
-      var mode = (stats.mode & 0100) ? 0777 : 0666;
+      var mode = (stats.mode & 0o100) ? 0o777 : 0o666;
       if (!options.transformContents) {
         copyFileHelper(fullFrom, fullTo, mode);
       } else {
@@ -542,7 +540,7 @@ files.findPathsWithRegex = function (dir, regex, options) {
 // have to exist. Treats symbolic links transparently (copies the contents, not
 // the link itself, and it's an error if the link doesn't point to a file).
 files.copyFile = function (from, to) {
-  files.mkdir_p(files.pathDirname(files.pathResolve(to)), 0755);
+  files.mkdir_p(files.pathDirname(files.pathResolve(to)), 0o755);
 
   var stats = files.stat(from);
   if (!stats.isFile()) {
@@ -554,7 +552,7 @@ files.copyFile = function (from, to) {
   // modified by umask.) We don't copy the mode *directly* because this function
   // is used by 'meteor create' which is copying from the read-only tools tree
   // into a writable app.
-  var mode = (stats.mode & 0100) ? 0777 : 0666;
+  var mode = (stats.mode & 0o100) ? 0o777 : 0o666;
 
   copyFileHelper(from, to, mode);
 };
@@ -605,7 +603,7 @@ files.mkdtemp = function (prefix) {
       var dirPath = files.pathJoin(
         tmpDir, prefix + (Math.random() * 0x100000000 + 1).toString(36));
       try {
-        files.mkdir(dirPath, 0700);
+        files.mkdir(dirPath, 0o700);
         return dirPath;
       } catch (err) {
         tries--;
@@ -766,7 +764,7 @@ files.createTarGzStream = function (dirPath, options) {
       // setting it in an 'entry' handler is the same strategy that npm
       // does, so we do that here too.
       if (entry.type === "Directory") {
-        entry.mode = (entry.mode || entry.props.mode) | 0500;
+        entry.mode = (entry.mode || entry.props.mode) | 0o500;
         entry.props.mode = entry.mode;
       }
 
@@ -858,10 +856,9 @@ files.symlinkOverSync = function (linkText, file) {
 // with the traditional POSIX execl(2).
 //
 // XXX 'files' is not the ideal place for this but it'll do for now
-files.run = function (command /*, arguments */) {
+files.run = function (command, ...args) {
   var Future = require('fibers/future');
   var future = new Future;
-  var args = _.toArray(arguments).slice(1);
 
   var child_process = require("child_process");
   child_process.execFile(
@@ -875,8 +872,7 @@ files.run = function (command /*, arguments */) {
   return future.wait();
 };
 
-files.runGitInCheckout = function (/* arguments */) {
-  var args = _.toArray(arguments);
+files.runGitInCheckout = function (...args) {
   args.unshift(
     'git', '--git-dir=' +
     files.convertToOSPath(files.pathJoin(files.getCurrentToolsDir(), '.git')));
@@ -1319,13 +1315,7 @@ function wrapFsFunc(fsFuncName, pathArgIndices, options) {
   var fsFunc = fs[fsFuncName];
   var fsFuncSync = fs[fsFuncName + "Sync"];
 
-  function wrapper() {
-    var argc = arguments.length;
-    var args = new Array(argc);
-    for (var i = 0; i < argc; ++i) {
-      args[i] = arguments[i];
-    }
-
+  function wrapper(...args) {
     for (var j = pathArgIndices.length - 1; j >= 0; --j) {
       i = pathArgIndices[j];
       args[i] = files.convertToOSPath(args[i]);
@@ -1429,38 +1419,33 @@ wrapFsFunc("symlink", [0, 1]);
 wrapFsFunc("readlink", [0]);
 
 // These don't need to be Fiberized
-files.createReadStream = function () {
-  var args = _.toArray(arguments);
+files.createReadStream = function (...args) {
   args[0] = files.convertToOSPath(args[0]);
-  return fs.createReadStream.apply(fs, args);
+  return fs.createReadStream(...args);
 };
 
-files.createWriteStream = function () {
-  var args = _.toArray(arguments);
+files.createWriteStream = function (...args) {
   args[0] = files.convertToOSPath(args[0]);
-  return fs.createWriteStream.apply(fs, args);
+  return fs.createWriteStream(...args);
 };
 
-files.watchFile = function () {
-  var args = _.toArray(arguments);
+files.watchFile = function (...args) {
   args[0] = files.convertToOSPath(args[0]);
-  return fs.watchFile.apply(fs, args);
+  return fs.watchFile(...args);
 };
 
-files.unwatchFile = function () {
-  var args = _.toArray(arguments);
+files.unwatchFile = function (...args) {
   args[0] = files.convertToOSPath(args[0]);
-  return fs.unwatchFile.apply(fs, args);
+  return fs.unwatchFile(...args);
 };
 
 // wrap pathwatcher because it works with file system paths
 // XXX we don't currently convert the path argument passed to the watch
 //     callback, but we currently don't use the argument either
-files.pathwatcherWatch = function () {
-  var args = _.toArray(arguments);
+files.pathwatcherWatch = function (...args) {
   args[0] = files.convertToOSPath(args[0]);
   // don't import pathwatcher until the moment we actually need it
   // pathwatcher has a record of keeping some global state
   var pathwatcher = require('pathwatcher');
-  return pathwatcher.watch.apply(pathwatcher, args);
+  return require("pathwatcher").watch(...args);
 };